requirements.txt
websocket-client
fastapi
uvicorn
pydantic
aiofiles
python-dotenv
aiohttp
pyaudio
pydub
openai
webrtcvad
websockets